- Сделан ThemeSwitcherService.kt
